src/main/java/com/orderinn/companyManagement/Api/CompanyController.java
package com.orderinn.companyManagement.Api;

import com.orderinn.companyManagement.Business.CompanyService;
import com.orderinn.companyManagement.Model.Company;
import com.orderinn.companyManagement.Model.User;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/api/company")
public class CompanyController {

    private final CompanyService companyService;

    @Autowired
    public CompanyController(CompanyService companyService) {
        this.companyService = companyService;
    }

    @GetMapping("/all")
    public ResponseEntity<List<Company>> getAllCompanies(){
        List<Company> companies = companyService.getAllCompanies();
        return new ResponseEntity<>(companies, HttpStatus.OK);
    }

    @GetMapping(path = "/{id}")
    public ResponseEntity<Company> getCompanyById(@PathVariable("id") Long Id){
        Company company;
        try{
            company = companyService.getCompanyById(Id);
            return new ResponseEntity<>(company, HttpStatus.OK);
        }catch (IllegalArgumentException e){
            e.printStackTrace();
            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
        }
    }

    @GetMapping("/employees/{id}")
    public ResponseEntity<List<User>> getEmployees(@PathVariable("id") Long Id){
        List<User> employees;
        try{
            employees = companyService.getEmployees(Id);
            return new ResponseEntity<>(employees, HttpStatus.OK);
        }catch (IllegalArgumentException e){
            e.printStackTrace();
            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
        }
    }

    @PostMapping("/new")
    public ResponseEntity<Company> addNewCompany(@RequestBody Company company){
        Company returnedCompany;
        try{
            returnedCompany = companyService.saveCompany(company);
            return new ResponseEntity<>(returnedCompany, HttpStatus.OK);
        }catch (IllegalArgumentException e){
            e.printStackTrace();
            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
        }
    }
}
